Make create_objects() a virtual function of BC_WindowBase.
This function was previously mostly non-virtual and a mix of int and void
was used as the return value. This makes the function virtual because in
the next change (that is about a code simplification in the plugin settings
windows) this will be useful. But being virtual means that the return
value must be the same in all derived classes. Therefore, all instances
of create_objects() are adjusted to return void (the int return value was
never used anyway).

This change is taken from Cinelerra 4.