Rework build information in the About box.
At configure time, the option --with-buildinfo can be used to insert the
git reversion in the About box (and the startup blurb). This used to offer
also to include the SVN revision number, but since we do not use SVN
anymore, this option is now removed.

Previously, the Makefile constructed the entire text that was inserted in
the code. This was changed so that the Makefile only creates macros for the
variable parts, but does not define how the parts are assembled. This way
it is easier to update the copyright statement.

The way in which the configure determined the possible options used some
temporary variables, but this can also be expressed with some not too
complicated case statements. (One gotcha is, however, that the unmatched
parenthesis that introduces a case arm must be quoted so that the autoconf
macro does not see it.)