Move the definition of class PluginWindow to pluginclient.h.
There is already a class definition of PluginClientWindow. This definition
should have been removed by 5ec21d77611c (Removed now unused class
PluginClientWindow), which removed only the member definitions. The class
exists in upstream Cinelerra 4 and is used as a base class for plugins,
just like our own PluginWindow class. To simplify merging Cinelerra 4,
rename PluginWindow to PluginClientWindow. But as a preparatory step, move
the class definition to its new place before we rename it in the next step.