Revert "r264: Fixes for faster playback of mp3 and vorbis, when output freq != input freq"
This reverts commit 0f5c610be664e83089b5f8cd9abe1333f459c222, which is a
speed optimization. When we merge Cinelerra 4.1 later, we would have to
duplicate a lot of changes made in the Resample class into the
Resample_float class. Since speed, particularly when it comes to audio
processing, should not be of a major concern today anymore, we take the
simple route and remove the optimization.

Two things to note are:

- The original change snuck in some changes into filempeg.C that are not
  about the topic of the change. We do not attempt to revert these changes.

- The original changes to filevorbis.* have already been reverted by the
  earlier merge of Cinelerra 4.