Factor formating of the master frame rate text into a helper function.
The text of the control is in a macro to avoid duplication in the two
sites where the text is needed. However, in order to allow translation,
the text is still duplicated into a #if 0 ... #endif section, which is
skipped by the compiler, but can be seen by xgettext.

Moreover, we will inherit an update from upstream that touches this
macro, but in a way that makes the text untranslatable. Forestall the
change with a clean solution.