Do not assume that the plugin index file is writable.
The new Cinelerra keeps a list of plugins in a file. It attempts to read
the file on startup, and if it is not present or the wrong version, it
attempts to write it. When Cinelerra is installed in a system location,
such as /usr/local, but is not run as root, the index file cannot be
written. This case was not taken into account.