Write converted image to stdout instead of to the current directory.
In a later change, we will invoke pngtoh in parallel from directory plugins
in parallel to convert */picon.png files. Obviously, it is detrimental to
write the result to the same file name picon_png.h (without the
subdirectory name!) in parallel.

To prepare for this use case, have pngtoh write the result to stdout and
require the caller to redirect stdout to the desired result file. Remove
the loop over the arguments because no user requires it.