Catch incorrect string literal suffixes.
The most likely source of incompatibility beginning with C++11 is when a
string is juxtaposed to a macro with no space between them, e.g.

   "The file has %"PRIu64" bytes."

Even though PRIu64 is a macro, it would be parsed as a user-defined literal
in combination with the string in front of it.

Switch GCC's warning about this case to an error and fix all occurrences
in the code base.